split bsp analyze/info/scale stuff from q3map2
