+++ tools/quake3/q3map2/bsp_scale.c
@@ -1,0 +1,143 @@
+/* -------------------------------------------------------------------------------
+
+   Copyright (C) 1999-2007 id Software, Inc. and contributors.
+   For a list of contributors, see the accompanying CONTRIBUTORS file.
+
+   This file is part of GtkRadiant.
+
+   GtkRadiant is free software; you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation; either version 2 of the License, or
+   (at your option) any later version.
+
+   GtkRadiant is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with GtkRadiant; if not, write to the Free Software
+   Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+
+   -------------------------------------------------------------------------------
+
+   This code has been altered significantly from its original form, to support
+   several games based on the Quake III Arena engine, in the form of "Q3Map2."
+
+   ------------------------------------------------------------------------------- */
+
+
+
+/* dependencies */
+#include "q3map2.h"
+
+
+
+/*
+   ScaleBSPMain()
+   amaze and confuse your enemies with wierd scaled maps!
+ */
+
+int ScaleBSPMain( int argc, char **argv ){
+	int i;
+	float f, scale;
+	vec3_t vec;
+	char str[ 1024 ];
+
+
+	/* arg checking */
+	if ( argc < 2 ) {
+		Sys_Printf( "Usage: q3map -scale <value> [-v] <mapname>\n" );
+		return 0;
+	}
+
+	/* get scale */
+	scale = atof( argv[ argc - 2 ] );
+	if ( scale == 0.0f ) {
+		Sys_Printf( "Usage: q3map -scale <value> [-v] <mapname>\n" );
+		Sys_Printf( "Non-zero scale value required.\n" );
+		return 0;
+	}
+
+	/* do some path mangling */
+	strcpy( source, ExpandArg( argv[ argc - 1 ] ) );
+	StripExtension( source );
+	DefaultExtension( source, ".bsp" );
+
+	/* load the bsp */
+	Sys_Printf( "Loading %s\n", source );
+	LoadBSPFile( source );
+	ParseEntities();
+
+	/* note it */
+	Sys_Printf( "--- ScaleBSP ---\n" );
+	Sys_FPrintf( SYS_VRB, "%9d entities\n", numEntities );
+
+	/* scale entity keys */
+	for ( i = 0; i < numBSPEntities && i < numEntities; i++ )
+	{
+		/* scale origin */
+		GetVectorForKey( &entities[ i ], "origin", vec );
+		if ( ( vec[ 0 ] + vec[ 1 ] + vec[ 2 ] ) ) {
+			VectorScale( vec, scale, vec );
+			sprintf( str, "%f %f %f", vec[ 0 ], vec[ 1 ], vec[ 2 ] );
+			SetKeyValue( &entities[ i ], "origin", str );
+		}
+
+		/* scale door lip */
+		f = FloatForKey( &entities[ i ], "lip" );
+		if ( f ) {
+			f *= scale;
+			sprintf( str, "%f", f );
+			SetKeyValue( &entities[ i ], "lip", str );
+		}
+	}
+
+	/* scale models */
+	for ( i = 0; i < numBSPModels; i++ )
+	{
+		VectorScale( bspModels[ i ].mins, scale, bspModels[ i ].mins );
+		VectorScale( bspModels[ i ].maxs, scale, bspModels[ i ].maxs );
+	}
+
+	/* scale nodes */
+	for ( i = 0; i < numBSPNodes; i++ )
+	{
+		VectorScale( bspNodes[ i ].mins, scale, bspNodes[ i ].mins );
+		VectorScale( bspNodes[ i ].maxs, scale, bspNodes[ i ].maxs );
+	}
+
+	/* scale leafs */
+	for ( i = 0; i < numBSPLeafs; i++ )
+	{
+		VectorScale( bspLeafs[ i ].mins, scale, bspLeafs[ i ].mins );
+		VectorScale( bspLeafs[ i ].maxs, scale, bspLeafs[ i ].maxs );
+	}
+
+	/* scale drawverts */
+	for ( i = 0; i < numBSPDrawVerts; i++ )
+		VectorScale( bspDrawVerts[ i ].xyz, scale, bspDrawVerts[ i ].xyz );
+
+	/* scale planes */
+	for ( i = 0; i < numBSPPlanes; i++ )
+		bspPlanes[ i ].dist *= scale;
+
+	/* scale gridsize */
+	GetVectorForKey( &entities[ 0 ], "gridsize", vec );
+	if ( ( vec[ 0 ] + vec[ 1 ] + vec[ 2 ] ) == 0.0f ) {
+		VectorCopy( gridSize, vec );
+	}
+	VectorScale( vec, scale, vec );
+	sprintf( str, "%f %f %f", vec[ 0 ], vec[ 1 ], vec[ 2 ] );
+	SetKeyValue( &entities[ 0 ], "gridsize", str );
+
+	/* write the bsp */
+	UnparseEntities();
+	StripExtension( source );
+	DefaultExtension( source, "_s.bsp" );
+	Sys_Printf( "Writing %s\n", source );
+	WriteBSPFile( source );
+
+	/* return to sender */
+	return 0;
+}
