+++ tools/quake3/q3map2/q3map2.vcxproj
@@ -174,6 +174,9 @@
     <ClCompile Include="brush.c" />
     <ClCompile Include="brush_primit.c" />
     <ClCompile Include="bsp.c" />
+    <ClCompile Include="bsp_analyze.c" />
+    <ClCompile Include="bsp_scale.c" />
+    <ClCompile Include="bsp_info.c" />
     <ClCompile Include="decals.c" />
     <ClCompile Include="facebsp.c" />
     <ClCompile Include="fog.c" />
